requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.22
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
python-slugify==8.0.1
email-validator==2.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.1.0 